+++ core/src/main/java/com/sequenceiq/cloudbreak/core/flow2/cluster/provision/ClusterCreationState.java
@@ -6,6 +6,8 @@ import com.sequenceiq.cloudbreak.core.flow2.FlowState;
 public enum ClusterCreationState implements FlowState {
     INIT_STATE,
     CLUSTER_CREATION_FAILED_STATE,
+    BOOTSTRAPPING_MACHINES_STATE,
+    COLLECTING_HOST_METADATA_STATE,
     STARTING_AMBARI_SERVICES_STATE,
     STARTING_AMBARI_STATE,
     INSTALLING_CLUSTER_STATE,
