+++ core/src/main/java/com/sequenceiq/cloudbreak/core/flow2/stack/Msg.java
@@ -35,6 +35,7 @@ public enum Msg {
     STACK_INFRASTRUCTURE_STOP_FAILED("stack.infrastructure.stop.failed"),
     STACK_INFRASTRUCTURE_ROLLBACK_MESSAGE("stack.infrastructure.create.rollback"),
     STACK_SYNC_INSTANCE_STATUS_COULDNT_DETERMINE("stack.sync.instance.status.couldnt.determine"),
+    FLOW_STACK_PROVISIONED_BILLING("flow.stack.provisioned.billing"),
     FLOW_STACK_PROVISIONED("flow.stack.provisioned"),
     FLOW_STACK_METADATA_COLLECTED("stack.metadata.collected"),
     CLUSTER_DELETE_COMPLETED("ambari.cluster.delete.completed"),
