+++ core/src/main/java/com/sequenceiq/cloudbreak/core/flow2/stack/provision/StackCreationState.java
@@ -14,8 +14,6 @@ public enum StackCreationState implements FlowState {
     PROVISIONING_FINISHED_STATE,
     COLLECTMETADATA_STATE,
     TLS_SETUP_STATE,
-    BOOTSTRAPING_MACHINES_STATE,
-    COLLECTING_HOST_METADATA_STATE,
     STACK_CREATION_FINISHED_STATE,
     FINAL_STATE;
