+++ core/src/main/resources/messages/messages.properties
@@ -34,7 +34,8 @@ stack.infrastructure.delete.failed=Infrastructure termination failed. Reason: {0
 stack.infrastructure.start.failed=Infrastructure start failed. Reason: {0}
 stack.infrastructure.stop.failed=Infrastructure stop failed. Reason: {0}
 stack.metadata.collected=Infrastructure metadata collection finished
-flow.stack.provisioned=Billing started, Infrastructure successfully provisioned
+flow.stack.provisioned=Infrastructure successfully provisioned
+flow.stack.provisioned.billing=Billing started, Infrastructure successfully provisioned
 stack.infrastructure.create.rollback=Could not create {0} instance(s), node count is decreased on group {1}. Reason: {2}
 
 ambari.cluster.starting=Starting Ambari cluster
